requirements.txt
gitpython
myst-parser
pandas
pre-commit
pydata-sphinx-theme
sphinx>=7
sphinx-design
sphinx-notfound-page
